feature: Remove unneeded Mojos
Fixes #1313
