+++ doc/src/main/asciidoc/inc/goals/_goals.adoc
@@ -8,10 +8,6 @@ This plugin supports a rich set for providing a smooth Java developer experience
 
 * <<goals-develop, Development goals>> target help not only in deploying resource descriptors to the development cluster but also to manage the lifecycle of the development cluster as well.
 
-* <<goals-infra, Infrastructure goals>> are good for setting up your Kubernetes and OpenShift development environment as well for adding this plugin to your `pom.xml`.
-
-* <<goals-internal, Internal goals>> are used by fabric8 project to maintain meta data of the supported applications but might be useful for other uses cases, too.
-
 [[goals-build]]
 .Build Goals
 [cols="1,3"]
@@ -32,17 +28,6 @@ This plugin supports a rich set for providing a smooth Java developer experience
 
 | *<<fabric8:resource-apply>>*
 | Run `fabric8:resource` ➜ `fabric8:apply`
-
-| *<<fabric8:helm>>*
-| Create a https://github.com/kubernetes/helm/blob/master/docs/charts.md#the-chart-file-structure[Helm Chart]
-
-| *<<fabric8:app-catalog>>*
-| Generate an app catalog
-
-
-| *<<fabric8:distro>>*
-| Generate an archive of Kubernetes and OpenShift templates
-
 |===
 
 [[goals-develop]]
@@ -51,68 +36,20 @@ This plugin supports a rich set for providing a smooth Java developer experience
 |===
 | Goal | Description
 
-| *<<fabric8:run>>*
-| Run a complete development workflow cycle `fabric8:resource` ➜ `fabric8:build` ➜ `fabric8:apply` in the foreground.
-
 | *<<fabric8:deploy>>*
 | Deploy resources decriptors to a cluster after creating them and building the app. Same as <<fabric8:run>> except that it runs in the backgorund.
 
 | *<<fabric8:undeploy>>*
 | Undeploy and remove resources decriptors from a cluster.
 
-| *<<fabric8:start>>*
-| Start the application which has been previously deployed
-
-| *<<fabric8:stop>>*
-| Stop the application which has been previously deployed
-
 | *<<fabric8:watch>>*
 | Watch for file changes and perform rebuilds and redeployments
-
-| *<<fabric8:watch-spring-boot>>*
-| Watch for local code changes in Spring Boot apps and restart the container on the fly
 
 | *<<fabric8:log>>*
 | Show the logs of the running application
 
 | *<<fabric8:debug>>*
 | Enable remote debugging
-|===
-
-[[goals-infra]]
-.Infrastructure Goals
-[cols="1,3"]
-|===
-| Goal | Description
-
-| *<<fabric8:setup>>*
-| Add this plugin to a given `pom.xml`
-
-| *<<fabric8:cluster-start>>*
-| Start a development cluster
-
-| *<<fabric8:cluster-stop>>*
-| Stop a development cluster
-
-| *<<fabric8:install>>*
-| Install a development cluster (via gofabric8) along with client side tools (kubectl, oc)
-|===
-
-[[goals-internal]]
-.Internal Goals
-
-[cols="1,3"]
-|===
-| Goal | Description
-
-| *<<fabric8:import>>*
-| Import the current project into the fabric8 console
-
-| *<<fabric8:helm-index>>*
-| Scan a Maven repository and create a Helm index
-
-| *<<fabric8:manifest-index>>*
-| Scan a Maven index and create a Manifest index
 |===
 
 Depending on whether the OpenShift or Kubernetes operational mode is used, the workflow and the performed actions differs :
@@ -149,5 +86,3 @@ a| `fabric8:deploy`
 
 include::build/_goals.adoc[]
 include::develop/_goals.adoc[]
-include::infra/_goals.adoc[]
-include::internal/_goals.adoc[]
