--- doc/src/main/asciidoc/inc/goals/develop/_fabric8-start.adoc
@@ -1,32 +1,0 @@
-
-[[fabric8:start]]
-== *fabric8:start*
-
-This goal starts the app that you deployed via the <<fabric8:deploy>> goal and then subsequently stopped via <<fabric8:stop>>.
-
-[source, sh]
-----
-mvn fabric8:start
-----
-
-This goal iterates through all scaleable resources generated by the <<fabric8:resource>> goal and scales them to a _replica count_ of`1` so that by the time this goal completes there should be a single running pod of this app.
-
-=== Scaling
-
-You can also use the `fabric8:start` goal to scale the number of pods running your application via passing the `fabric8.replicas` parameter:
-
-[source, sh]
-----
-mvn fabric8:start  -Dfabric8.replicas=2
-----
-
-After the above you should see 2 pods running for your application.
-
-From the command line you can use the `kubectl` binary to watch pods come and go via the following:
-
-[source, sh]
-----
-kubectl get pod -w
-----
-
-The `-w` flag watches the resource and keeps updating the console as things change. Leaving off the `-w` just lists the current pods only then `kubect` terminates.
