--- doc/src/main/asciidoc/inc/goals/develop/_fabric8-stop.adoc
@@ -1,17 +1,0 @@
-
-[[fabric8:stop]]
-== *fabric8:stop*
-
-This goal stops the app that you deployed via the <<fabric8:deploy>> goal.
-
-[source, sh]
-----
-mvn fabric8:stop
-----
-
-This goal iterates through all scaleable resources generated by the <<fabric8:resource>> goal and scales them to a _replica count_ of `0` so that by the time this goal completes there are no running pods for this app.
-
-### Removing all resources
-
-Note that if you wish to completely remove all resources for your application then use <<fabric8:undeploy>> goal
-
