--- doc/src/main/asciidoc/inc/goals/infra/_fabric8-cluster-start.adoc
@@ -1,92 +1,0 @@
-
-[[fabric8:cluster-start]]
-== *fabric8:cluster-start*
-
-This goal will start a local kubernetes cluster for local development.
-
-### Prerequisites
-
-Please use a recent distribution of http://maven.apache.org/download.cgi[Apache Maven] at least 3.3.x or later. We use `3.3.9` here and it works well!
-
-Depending on your platform you also need to install the following drivers:
-
-* Windows users will need to run this command as Adminstrator and will need to https://msdn.microsoft.com/en-us/virtualization/hyperv_on_windows/quick_start/walkthrough_install[enable Hyper-V on Windows 10] or https://blogs.technet.microsoft.com/schadinio/2010/07/09/installing-hyper-v-manager-on-windows-7/[Windows 7].
-* OS X users  will need to https://github.com/kubernetes/minikube/blob/master/DRIVERS.md#xhyve-driver[install the xhyve driver] which we try to automatically install via `brew` but you may want to install it just in case ;)
-* Linux will need to https://github.com/kubernetes/minikube/blob/master/DRIVERS.md#kvm-driver[install the kvm driver]
-
-
-### Starting the cluster
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-start
-----
-
-This will internally invoke the <<fabric8:install>> goal to ensure that all the required binaries are installed (like https://github.com/fabric8io/gofabric8/[gofabric8]
- and for kubernetes: http://kubernetes.io/docs/user-guide/kubectl-overview/[kubectl] and https://github.com/kubernetes/minikube[minikube] or for OpenShift: https://docs.openshift.com/enterprise/latest/cli_reference/basic_cli_operations.html[oc] and https://github.com/jimmidyson/minishift[minishift])
-
-By default the binaries are installed in `~/.fabric8/bin`
-
-=== Using OpenShift
-
-By default `fabric8:cluster-start` will use `minikube` to create a local single node kubernetes cluster. To specify OpenShift use:
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-start -Dfabric8.cluster.kind=openshift
-----
-
-This will then use `minishift` instead to create a single node local OpenShift cluster.
-
-=== VM drivers
-
-By default the VM drivers used will be `hyperv` on Windows, `xhyve` on OS X and `kvm` on Linux.
-
-If you wish to switch to a different VM driver you can specify the `fabric8.cluster.driver` property. For example if you have installed https://www.virtualbox.org/wiki/Downloads[VirtualBox] and wish to use that then type:
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-start -Pfmp-snapshot -Dfabric8.cluster.driver=virtualbox
-----
-
-Note that we highly recommend using the default VM drivers (`hyperv` on Windows, `xhyve` on OS X and `kvm` on Linux) as they tend to work better and use less resources on your laptop than the alternatives.
-
-
-=== Configure apps
-
-By default the cluster contains only the http://fabric8.io/guide/console.html[fabric8 developer console] as often developers laptops don't have lots of RAM.
-
-If you want to deploy the full fabric8 platform (with Nexus, Jenkins, Gogs, JBoss Forge etc) then use the following command:
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-start -Dfabric8.cluster.app=platform
-----
-
-=== Configure cluster resources
-
-You can specify the number of CPUs or memory via additional parameters:
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-start -Dfabric8.cluster.cpus=2 -Dfabric8.cluster.memory=4096
-----
-
-The above configures `2` CPUs and `4Gb` of memory
-
-=== Stop
-
-You can stop the cluster at any time via <<fabric8:cluster-stop>>
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-stop
-----
-
-Once stopped you can restart again with all the images, resources and pods intact later on by running `fabric8:cluster-start` again
-
-[source,sh,subs="attributes"]
-----
-mvn fabric8:cluster-start
-----
-
