--- doc/src/main/asciidoc/inc/goals/infra/_goals.adoc
@@ -1,8 +1,0 @@
-
-[[goals-overview-setup]]
-# Infrastructure Goals
-
-include::_fabric8-setup.adoc[]
-include::_fabric8-cluster-start.adoc[]
-include::_fabric8-cluster-stop.adoc[]
-include::_fabric8-install.adoc[]
