--- doc/src/main/asciidoc/inc/goals/internal/_fabric8-helm-index.adoc
@@ -1,5 +1,0 @@
-
-[[fabric8:helm-index]]
-== *fabric8:helm-index*
-
-Generates a Manifest index file by querying a maven repository to find all the Kubernetes and OpenShift manifests available and their releases.
