--- doc/src/main/asciidoc/inc/goals/internal/_fabric8-manifest-index.adoc
@@ -1,3 +1,0 @@
-
-[[fabric8:manifest-index]]
-== *fabric8:manifest-index*
