--- doc/src/main/asciidoc/inc/goals/internal/_goals.adoc
@@ -1,6 +1,0 @@
-[[goals-overview-internal]]
-# Internal Goals
-
-include::_fabric8-import.adoc[]
-include::_fabric8-helm-index.adoc[]
-include::_fabric8-manifest-index.adoc[]
