--- doc/src/main/asciidoc/inc/goals/shortcut/_goals.adoc
@@ -1,6 +1,0 @@
-
-[[goals-overview-shortcut]]
-# Shortcut Goals
-
-include::_fabric8-run.adoc[]
-include::_fabric8-resource-apply.adoc[]
