+++ it/src/it/docker-compose/invoker.properties
@@ -14,6 +14,6 @@
 # permissions and limitations under the License.
 #
 
-invoker.goals.1=clean fabric8:install fabric8:resource
+invoker.goals.1=clean fabric8:resource
 invoker.mavenOpts=-Dfabric8.verbose -Dfabric8.mode=kubernetes -Dfabric8.install.batch.mode=true
 invoker.debug=false
