+++ it/src/it/vertx/pom.xml
@@ -76,7 +76,6 @@
             <id>fmp</id>
             <goals>
               <goal>resource</goal>
-              <goal>helm</goal>
               <goal>build</goal>
             </goals>
           </execution>
