--- plugin/src/test/java/io/fabric8/maven/plugin/HelmEscapeTest.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2016 Red Hat, Inc.
- *
- * Red Hat licenses this file to you under the Apache License, version
- * 2.0 (the "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied.  See the License for the specific language governing
- * permissions and limitations under the License.
- */
-package io.fabric8.maven.plugin;
-
-import org.junit.Test;
-
-import static io.fabric8.maven.plugin.mojo.build.HelmMojo.escapeYamlTemplate;
-import static org.junit.Assert.assertEquals;
-
-/**
- */
-public class HelmEscapeTest {
-
-    @Test
-    public void testHelmTemplateEscape() throws Exception {
-        assertEquals("abcd", escapeYamlTemplate("abcd"));
-        assertEquals("abc{de}f}", escapeYamlTemplate("abc{de}f}"));
-        assertEquals("abc{{\"{{\"}}de}f", escapeYamlTemplate("abc{{de}f"));
-        assertEquals("abc{{\"{{\"}}de}f{{\"}}\"}}", escapeYamlTemplate("abc{{de}f}}"));
-    }
-
-}
